# Backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Borrowss API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/V1/sub-categories": {
            "get": {
                "tags": [
                    "Masters"
                ],
                "summary": "Get list of sub-categories",
                "description": "Returns a list of sub-categories with their details.",
                "operationId": "getSubCategories",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "responseData": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "categories_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "TV"
                                                    },
                                                    "status": {
                                                        "type": "integer",
                                                        "example": 0
                                                    },
                                                    "deleted_at": {
                                                        "type": "string",
                                                        "example": null
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "02-07-2024"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "02-07-2024"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Sub Categories List"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
...
        "/api/V1/get-testimonials": {
            "get": {
                "tags": [
                    "Testimonials"
                ],
                "summary": "Get list of testimonials",
                "description": "Returns list of testimonials",
                "operationId": "getTestimonials",
                "parameters": [
                    {
                        "name": "X-Api-Key",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "responseData": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "user_name": {
                                                        "type": "string",
                                                        "example": "asdasdasdas"
                                                    },
                                                    "user_position": {
                                                        "type": "string",
                                                        "example": "asd"
                                                    },
                                                    "user_company": {
                                                        "type": "string",
                                                        "example": "asdad"
                                                    },
                                                    "testimonial": {
                                                        "type": "string",
                                                        "example": "asdasdad"
                                                    },
                                                    "photo": {
                                                        "type": "string",
                                                        "example": "http://192.168.1.149:8002/storage/testimonial_photos/0wAdKTOgw5d2YXgs81UmoFKpiPlxsVSdFVJftS39.jpg"
                                                    },
                                                    "deleted_at": {
                                                        "type": "string",
                                                        "example": null,
                                                        "nullable": true
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "16-07-2024"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "16-07-2024"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Testimonial List."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        }
    },
...
}